refactor: add phpstan and update code to adhere to level 5
- move and refactor Image.php from Libraries to Entities folder
- update some database field names
/ types
- update composer packages

+++ app/Libraries/Analytics/Database/Migrations/2017-12-01-160000_add_analytics_podcasts_by_region.php
@@ -36,11 +36,11 @@ class AddAnalyticsPodcastsByRegion extends Migration
                 'comment' => 'ISO 3166-2 code.',
             ],
             'latitude' => [
-                'type' => 'FLOAT',
+                'type' => 'DECIMAL',
                 'null' => true,
             ],
             'longitude' => [
-                'type' => 'FLOAT',
+                'type' => 'DECIMAL',
                 'null' => true,
             ],
             'hits' => [
